+++ examples/vanilla/CHANGELOG.md
@@ -1,5 +1,19 @@
 # example-vanilla
 
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies
+  - pinch-zoom-native@0.0.18
+
+## 0.0.18
+
+### Patch Changes
+
+- Updated dependencies
+  - pinch-zoom-native@0.0.17
+
 ## 0.0.17
 
 ### Patch Changes